requirements.txt
undetected-chromedriver
selenium
python-dotenv
pyyaml
schedule
requests 